refactor(nixl): fold pega pull helpers into pull modules
